+++ Assets/Scripts/HP/EntityHealthComponent.cs
@@ -23,7 +23,7 @@ public class EntityHealthComponent : HealthComponent
         entityHPBar.transform.position = transform.position + Vector3.up * height;
     }
 
-    public override void DeductHealth(float value)
+    public override void DeductHealth(int value)
     {
         health = Mathf.Clamp(health - value, 0, maxHealth);
