+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,7 +1,7 @@
 
 public interface IDamageable
 {
-    public void ApplyDamage(float damage);
+    public void ApplyDamage(int damage);
 
-    public void ApplyDamageOverTime(float damage, float duration);
+    public void ApplyDamageOverTime(int damage, float duration);
 }
